Migrate nyancat to inferno
